Refactor existing tests
